Move handling logic to route handlers
